+++ packages/ui/cascader/cascader.module.css
@@ -6,7 +6,7 @@
 .cos-cascader .cos-trigger {
     display: inline-flex;
     align-items: center;
-    padding: 0 var(--spacing-10);
+    padding: 0 1rem;
     border: 1px solid hsla(0, 0%, 12%, 0.1);
     height: var(--height-md);
     width: 20.6rem;
@@ -16,7 +16,7 @@
     border: 1px solid hsla(0, 0%, 12%, 1);
 }
 .cos-cascader .cos-popover-wrapper {
-    margin-top: var(--spacing-8);
+    margin-top: var(--padding-sm);
     position: absolute;
     z-index: 100;
     top: 100%;
@@ -29,6 +29,7 @@
 }
 .cos-trigger-input input {
     border: none;
+    background: transparent;
 }
 .cos-trigger-input input:focus {
     outline: none;
@@ -36,7 +37,7 @@
 .cos-cascader-items {
     height: 11rem;
     display: flex;
-    color: var(--color-light);
+    color: var(--color-white);
 }
 .cos-cascader-items .cos-left {
     height: 100%;
@@ -45,8 +46,8 @@
     overflow: auto;
     -webkit-overflow-scrolling: auto;
     background-color: hsl(240,8%,16%);
-    border-radius: var(--spacing-8);
-    margin-right: var(--spacing-2);
+    border-radius: var(--padding-sm);
+    margin-right: .2rem;
 }
 .cos-cascader-items .cos-left .cos-label {
     padding: .7rem 1rem .7rem 2rem;
@@ -80,6 +81,6 @@
     /* border-left: 2px solid #e6e6e6; */
     overflow: auto;
     -webkit-overflow-scrolling: auto;
-    border-radius: var(--spacing-8);
+    border-radius: var(--padding-sm);
     background-color: hsl(240,8%,16%);
 }
