+++ packages/vue/site/App.vue
@@ -17,7 +17,6 @@ import {
     Collapse,
     CollapseItem,
     Popover,
-    Cascader,
 } from 'cosmic-vue';
 import ArrowRight from '~icons/cosmic/arrow-right';
 import { ref } from 'vue';
@@ -30,6 +29,7 @@ import StoryGrid from './src/story-grid.vue';
 import ThemePicker from './src/theme-picker.vue';
 import StoryVariantUtilities from './src/story-variant-classes.vue';
 import StoryStep from './src/story-step.vue';
+import StoryCascader from './src/story-cascader.vue';
 
 const select = ref(null);
 
@@ -90,32 +90,7 @@ function log(msg: any) {
     // eslint-disable-next-line no-console
     console.log(msg);
 }
-const selected = ref([]);
-const source = ref([
-    {
-        name: '视觉搜索',
-        children: [
-            {
-                name: '识货',
-                children: [{ name: '通用版' }, { name: '春节' }, { name: '购物节' }],
-            },
-            {
-                name: '人脸识别',
-            },
-        ],
-    },
-    
-    {
-        name: '视频多模沉浸',
-    },
-    {
-        name: '语音搜索',
-    },
-]);
-function fn() {
-    // eslint-disable-next-line no-console
-    console.log('app....', selected.value);
-}
+
 </script>
 <template>
     <div>
@@ -304,12 +279,7 @@ function fn() {
                     collapse item {{ item }}
                 </collapse-item>
             </collapse>
-            <Cascader
-                v-model:source="source"
-                v-model:selected="selected"
-                popover-height="200px"
-                @update:selected="fn"
-            />
+            <story-cascader />
         </div>
     </div>
 </template>
